% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpivot.R
\name{rpivot}
\alias{rpivot}
\title{Create orb pivot widget}
\usage{
rpivot(data = NULL, fields, width = NULL, height = NULL)
}
\arguments{
\item{data}{a data object (currently supported objects are matrices, data
frames, and data.table}

\item{width}{the width of the pivot}

\item{height}{the height of the pivot}
}
\value{
A HTML widget object).
}
\description{
This function creates an orb widget using \pkg{htmlwidgets}. The
widget can be rendered on HTML pages generated from R Markdown, Shiny, or
other applications.
}
\examples{
inst/examples/pivot.R
}

