% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpivot.R
\name{rpivot}
\alias{rpivot}
\title{Create orb pivot widget}
\usage{
rpivot(dataSource, fields, displayMode = NULL, name = "pgridwidget",
  rows = NULL, columns = NULL, data = NULL, width = NULL,
  height = NULL, dataSettings = NULL, dataHeadersLocation = "columns",
  preFilters = NULL, toolbar = list(visible = TRUE), theme = "green",
  update = FALSE)
}
\arguments{
\item{dataSource}{a data object (currently supported objects are matrices, data
frames, and data.table}

\item{width}{the width of the pivot}

\item{height}{the height of the pivot}
}
\value{
A HTML widget object).
}
\description{
This function creates an orb widget using \pkg{htmlwidgets}. The
widget can be rendered on HTML pages generated from R Markdown, Shiny, or
other applications.
}
\examples{
inst/examples/pivot.R
}

